backend/requirements.txt
fastapi
uvicorn
pymongo
faiss-cpu
pypdf
sentence-transformers
python-dotenv
groq
python-multipart